Airflow/requirements.txt
apache-airflow-providers-snowflake==6.5.4